extracted TimeSortOptions component to its own file
