+++ blogproject/src/ViewDrafts.js
@@ -6,7 +6,7 @@ import AdvancedFormat from 'dayjs/plugin/advancedFormat'
 import AlertText from './AlertText'
 import SearchOptions from './SearchOptions'
 import DateSortOptions from './DateSortOptions'
-// import TimeSortOptions from './TimeSortOptions'
+import TimeSortOptions from './TimeSortOptions'
 import TitleSortOptions from './TitleSortOptions'
 
 
@@ -281,59 +281,6 @@ useEffect(()=>{
         </Container>
     )
 }
-
-const TimeSortOptions = (props)=>{
-
-    const handleTimeSort = (array, sortDirection)=>{
-        let newArray
-        switch(sortDirection){
-            case 'ascending':
-                newArray = array.sort((a,b)=>{
-                    return a.sortableTime - b.sortableTime
-                })    
-                
-            console.log('time ascent case', sortDirection, newArray);
-            break;
-            case 'descending':
-                newArray = array.sort((a,b)=>{
-                    return b.sortableTime - a.sortableTime
-                })      
-               
-            console.log('time descent case',  sortDirection, newArray);
-            break;
-        default:
-        console.log('sortclicked type has no value', sortDirection)
-          
-        }
-        props.setNewArray(newArray)
-        props.toggleSort(false)
-        }
-    return(
-        <Container style={{border: '2px solid red', width: '50%'}}>
-       <h3>By Time</h3>
-
-        <Form.Group>
-                    <Button
-                    onClick={()=>{
-                        handleTimeSort(props.array,'descending')
-                    }}
-                    >Oldest first</Button>
-                    <Button
-                    onClick={()=>{
-                        handleTimeSort(props.array,'ascending')
-                    }}
-                    >Newest first</Button>
-                    </Form.Group>
-        </Container>
-    )
-}
-
-
-
-
-
-
-
 
   
 export default ViewDrafts
